app/requirements.txt
flask
torch
transformers
pinecone-client
flask-cors